requirements.txt
# =============================================================================
# Forex Signal SaaS - Python Dependencies
# =============================================================================

# Machine Learning & Deep Learning
# Machine Learning & Deep Learning
tensorflow>=2.15.0
scikit-learn>=1.3.0

# Technical Analysis
pandas_ta>=0.3.14b0

# Market Data Sources
yfinance>=0.2.36
twelvedata>=1.2.13

# Environment & Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Dashboard & Visualization
streamlit>=1.31.0
plotly>=5.18.0
matplotlib>=3.8.0

# Telegram Bot Integration
python-telegram-bot>=20.7

# =============================================================================
# Core Data Processing (commonly needed)
# =============================================================================
pandas>=2.2.0
numpy>=1.26.0

# =============================================================================
# Development Tools (optional but recommended)
# =============================================================================
pytest>=8.0.0
black>=24.1.0
flake8>=7.0.0
